# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restaurant QR Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  /** @nullable */
  icon?: string | null;
  sortOrder: number;
  dishCount?: number;
  createdAt: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  icon?: string;
  sortOrder?: number;
}

export interface Dish {
  id: number;
  /** @nullable */
  categoryId?: number | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  modelGlbUrl?: string | null;
  isAvailable: boolean;
  sortOrder: number;
  createdAt: string;
}

export interface DishInput {
  /** @nullable */
  categoryId?: number | null;
  /** @minLength 1 */
  name: string;
  description?: string;
  price: string;
  imageUrl?: string;
  modelGlbUrl?: string;
  isAvailable?: boolean;
  sortOrder?: number;
}

export interface DishUpdate {
  /** @nullable */
  categoryId?: number | null;
  /** @minLength 1 */
  name?: string;
  description?: string;
  price?: string;
  imageUrl?: string;
  modelGlbUrl?: string;
  isAvailable?: boolean;
  sortOrder?: number;
}

export interface QrCode {
  id: number;
  /** @nullable */
  label?: string | null;
  targetUrl: string;
  /** @nullable */
  pngBase64?: string | null;
  createdAt: string;
}

export interface Stats {
  totalDishes: number;
  totalCategories: number;
  modelsCount: number;
}

export type ListDishesParams = {
category_id?: number;
};