# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant QR Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: categories
    description: Menu categories
  - name: dishes
    description: Menu dishes
  - name: qr
    description: QR code management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories with dish count
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /dishes:
    get:
      operationId: listDishes
      tags: [dishes]
      summary: List dishes, optionally filtered by category
      parameters:
        - name: category_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of dishes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dish"
    post:
      operationId: createDish
      tags: [dishes]
      summary: Create a dish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DishInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"

  /dishes/{id}:
    get:
      operationId: getDish
      tags: [dishes]
      summary: Get dish by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dish detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"
        "404":
          description: Not found
    put:
      operationId: updateDish
      tags: [dishes]
      summary: Update a dish
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DishUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"
        "404":
          description: Not found
    delete:
      operationId: deleteDish
      tags: [dishes]
      summary: Delete a dish
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /dishes/{id}/toggle:
    patch:
      operationId: toggleDishAvailability
      tags: [dishes]
      summary: Toggle dish availability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dish"
        "404":
          description: Not found

  /qr:
    get:
      operationId: getQrCode
      tags: [qr]
      summary: Get current QR code
      responses:
        "200":
          description: Current QR code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
        "404":
          description: No QR code yet

  /qr/generate:
    post:
      operationId: generateQrCode
      tags: [qr]
      summary: Generate a new QR code
      responses:
        "201":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"

  /stats:
    get:
      operationId: getStats
      tags: [categories]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Category:
      type: object
      required: [id, name, sortOrder, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: ["string", "null"]
        sortOrder:
          type: integer
        dishCount:
          type: integer
        createdAt:
          type: string

    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: string
        sortOrder:
          type: integer

    Dish:
      type: object
      required: [id, name, price, isAvailable, sortOrder, createdAt]
      properties:
        id:
          type: integer
        categoryId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        imageUrl:
          type: ["string", "null"]
        modelGlbUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string

    DishInput:
      type: object
      required: [name, price]
      properties:
        categoryId:
          type: ["integer", "null"]
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: string
        imageUrl:
          type: string
        modelGlbUrl:
          type: string
        isAvailable:
          type: boolean
        sortOrder:
          type: integer

    DishUpdate:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: string
        imageUrl:
          type: string
        modelGlbUrl:
          type: string
        isAvailable:
          type: boolean
        sortOrder:
          type: integer

    QrCode:
      type: object
      required: [id, targetUrl, createdAt]
      properties:
        id:
          type: integer
        label:
          type: ["string", "null"]
        targetUrl:
          type: string
        pngBase64:
          type: ["string", "null"]
        createdAt:
          type: string

    Stats:
      type: object
      required: [totalDishes, totalCategories, modelsCount]
      properties:
        totalDishes:
          type: integer
        totalCategories:
          type: integer
        modelsCount:
          type: integer