# lib/api-zod/src/generated/types/dish.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restaurant QR Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Dish {
  id: number;
  /** @nullable */
  categoryId?: number | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  modelGlbUrl?: string | null;
  isAvailable: boolean;
  sortOrder: number;
  createdAt: string;
}