# lib/api-zod/src/generated/types/dishUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restaurant QR Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DishUpdate {
  /** @nullable */
  categoryId?: number | null;
  /** @minLength 1 */
  name?: string;
  description?: string;
  price?: string;
  imageUrl?: string;
  modelGlbUrl?: string;
  isAvailable?: boolean;
  sortOrder?: number;
}