# lib/api-zod/src/generated/types/qrCode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restaurant QR Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface QrCode {
  id: number;
  /** @nullable */
  label?: string | null;
  targetUrl: string;
  /** @nullable */
  pngBase64?: string | null;
  createdAt: string;
}